+++ src/Cool/Gpu/internal/RenderTarget_ImplOpenGL.h
@@ -20,7 +20,7 @@ public:
     ImageSize           size() const { return _texture.size(); }
     ImageData           download_pixels() const;
     // const Texture&   texture() const { return _texture.; }
-    ImTextureID imgui_texture_id() const { return reinterpret_cast<ImTextureID>(_texture.textureID()); }
+    ImTextureID imgui_texture_id() const { return reinterpret_cast<ImTextureID>(static_cast<uint64_t>(_texture.textureID())); } // Double-cast to fix a warning : first we convert to the correct size (uint32_t -> uint64_t) then from integral type to pointer type (uint64_t -> ImTextureID)
 
     void imgui_window();
